refactor: move uniswap test to liquidity manager test
